Теперь код работает. Снес файл special_functions.js потому что импорт функции ломает код в браузере. в другой раз будем экспериментировать.
